requirements.txt
django==3.0
django-allauth==0.42
django-cors-headers==3.4.0
django-cors-middleware==1.5
django-countries==6.1
django-exclusivebooleanfield==0.3.0 # Not updated for Django 3
django-oauth-toolkit==1.3.2
django-widget-tweaks==1.4.8 # Not updated for Django 3
djangorestframework==3.11
humanfriendly
humanize
pendulum==2.0.5
Pillow
psycopg2-binary
Pygments
python-dateutil==2.8.1
requests
